# docs/swagger.yaml
basePath: /
definitions:
...
  http.ErrorResponse:
    properties:
      message:
        description: Error message
        type: string
    type: object
  http.StatusResponse:
    properties:
      status:
        description: Status message
        type: string
    type: object
...
  rule.CreateRuleInput:
    properties:
      ptype:
        description: Type
        type: string
      v0:
        description: Role name
        type: string
      v1:
        description: Recourse name
        type: string
      v2:
        description: REST verb
        type: string
      v3:
        description: Permission status
        type: string
      v4:
        description: Empty
        type: string
      v5:
        description: Empty
        type: string
    required:
    - ptype
    - v0
    - v1
    - v2
    - v3
    type: object
  rule.Rule:
    properties:
      id:
        description: Rule ID
        type: string
      ptype:
        description: Type
        type: string
      v0:
        description: Role name
        type: string
      v1:
        description: Recourse name
        type: string
      v2:
        description: REST verb
        type: string
      v3:
        description: Permission status
        type: string
      v4:
        description: Empty
        type: string
      v5:
        description: Empty
        type: string
    required:
    - ptype
    - v0
    - v1
    - v2
    - v3
    type: object
  rule.UpdateRuleInput:
    properties:
      id:
        description: Rule ID
        type: string
      ptype:
        description: Type
        type: string
      v0:
        description: Role name
        type: string
      v1:
        description: Recourse name
        type: string
      v2:
        description: REST verb
        type: string
      v3:
        description: Permission status
        type: string
      v4:
        description: Empty
        type: string
      v5:
        description: Empty
        type: string
    type: object
  specification.CreateSpecificationInput:
    properties:
      descriptionen:
        description: Description English
        type: string
      descriptionru:
        description: Description Russian
        type: string
      descriptiontm:
        description: Description Turkmen
        type: string
      descriptiontr:
        description: Description Turkish
        type: string
      item:
        description: Item ID
        type: string
      nameen:
        description: Name English
        type: string
      nameru:
        description: Name Russian
        type: string
      nametm:
        description: Name Turkmen
        type: string
      nametr:
        description: Name Turkish
        type: string
      organization:
        description: Organization ID
        type: string
      value:
        description: Value
        type: string
    required:
    - item
    - nameen
    - nameru
    - nametm
    - nametr
    - organization
    - value
    type: object
  specification.Specification:
    properties:
      descriptionen:
        description: Description English
        type: string
      descriptionru:
        description: Description Russian
        type: string
      descriptiontm:
        description: Description Turkmen
        type: string
      descriptiontr:
        description: Description Turkish
        type: string
      id:
        description: Specification ID
        type: string
      item:
        description: Item ID
        type: string
      nameen:
        description: Name English
        type: string
      nameru:
        description: Name Russian
        type: string
      nametm:
        description: Name Turkmen
        type: string
      nametr:
        description: Name Turkish
        type: string
      organization:
        description: Organization ID
        type: string
      value:
        description: Value
        type: string
    required:
    - item
    - nameen
    - nameru
    - nametm
    - nametr
    - organization
    - value
    type: object
  specification.UpdateSpecificationInput:
    properties:
      descriptionen:
        description: Description English
        type: string
      descriptionru:
        description: Description Russian
        type: string
      descriptiontm:
        description: Description Turkmen
        type: string
      descriptiontr:
        description: Description Turkish
        type: string
      id:
        description: Specification ID
        type: string
      item:
        description: Item ID
        type: string
      nameen:
        description: Name English
        type: string
      nameru:
        description: Name Russian
        type: string
      nametm:
        description: Name Turkmen
        type: string
      nametr:
        description: Name Turkish
        type: string
      organization:
        description: Organization ID
        type: string
      value:
        description: Value
        type: string
    type: object
...
info:
  contact:
    email: evgeniydammer@gmail.com
    name: Evgeniy Dammer
  description: marketplace API service
  title: marketplace API
  version: "1.0"
paths:
...
  /api/v1/rules/:
    get:
      consumes:
      - application/json
      description: Get all rules method.
      produces:
      - application/json
      responses:
        "200":
          description: Rule List
          schema:
            items:
              $ref: '#/definitions/rule.Rule'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all rules method.
      tags:
      - rules
    patch:
      consumes:
      - application/json
      description: Update rule method.
      parameters:
      - description: Rule data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/rule.UpdateRuleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Update rule method.
      tags:
      - rules
    post:
      consumes:
      - application/json
      description: Create rule method.
      parameters:
      - description: Rule data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/rule.CreateRuleInput'
      produces:
      - application/json
      responses:
        "200":
          description: Rule ID
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Create rule method.
      tags:
      - rules
  /api/v1/rules/{id}:
    delete:
      consumes:
      - application/json
      description: Delete rule method.
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete rule method.
      tags:
      - rules
    get:
      consumes:
      - application/json
      description: Get rule by id method.
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Rule data
          schema:
            $ref: '#/definitions/rule.Rule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Get rule by id method.
      tags:
      - rules
  /api/v1/specifications/:
    patch:
      consumes:
      - application/json
      description: Update specification method.
      parameters:
      - description: Specification data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/specification.UpdateSpecificationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Update specification method.
      tags:
      - specifications
    post:
      consumes:
      - application/json
      description: Create specification method.
      parameters:
      - description: Specification data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/specification.CreateSpecificationInput'
      produces:
      - application/json
      responses:
        "200":
          description: Specification ID
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Create specification method.
      tags:
      - specifications
  /api/v1/specifications/{org_id}:
    get:
      consumes:
      - application/json
      description: Get all specifications method.
      parameters:
      - description: Specification ID
        in: path
        name: org_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Specification List
          schema:
            items:
              $ref: '#/definitions/specification.Specification'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all specifications method.
      tags:
      - specifications
  /api/v1/specifications/{org_id}/{id}:
    delete:
      consumes:
      - application/json
      description: Delete specification method.
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      - description: Specification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete specification method.
      tags:
      - specifications
    get:
      consumes:
      - application/json
      description: Get specification by id method.
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      - description: Specification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Specification data
          schema:
            $ref: '#/definitions/specification.Specification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - Bearer: []
      summary: Get specification by id method.
      tags:
      - specifications
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"